+++ packages/core/src/api/signOut.ts
@@ -1,5 +1,5 @@
 import { HeadersBuilder } from "@aura-stack/router"
-import { isAuraAuthError } from "@/shared/unstable_error.ts"
+import { isAuraAuthError } from "@/shared/errors.ts"
 import { createRedirectTo, getBaseURL, getOriginURL } from "@/actions/signIn/authorization.ts"
 import type { FunctionAPIContext, SignOutAPIOptions, SignOutAPIReturn } from "@/@types/index.ts"
