+++ packages/core/src/cookie.ts
@@ -1,5 +1,5 @@
 import { env } from "@/shared/env.ts"
-import { AuraAuthError } from "@/shared/unstable_error.ts"
+import { AuraAuthError } from "@/shared/errors.ts"
 import { parse, parseSetCookie, serialize, type SerializeOptions } from "@aura-stack/router/cookie"
 import type { CookieStoreConfig, CookieConfig, InternalLogger } from "@/@types/index.ts"
