+++ packages/core/src/jose.ts
@@ -28,7 +28,7 @@ import {
 import { importPEMKeyPair } from "@/shared/crypto.ts"
 export { encoder, getRandomBytes, getSubtleCrypto } from "@aura-stack/jose/crypto"
 import type { User, SessionConfig, JWTKey, AsymmetricKeyPairFromEnv } from "@/@types/index.ts"
-import { AuraAuthError } from "./shared/unstable_error.ts"
+import { AuraAuthError } from "./shared/errors.ts"
 
 const getJWTConfig = (config?: SessionConfig) => {
     return config?.jwt
