+++ packages/hono/package.json
