+++ packages/react/package.json
